requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
diff-match-patch==20241021
Django==4.2.19
django-autoslug==1.9.9
django-ckeditor==6.7.2
django-filter==25.1
django-import-export==4.3.7
django-js-asset==2.2.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
phonenumbers==8.13.55
pillow==11.1.0
psycopg==3.2.5
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
zipp==3.21.0
